refactor(db): :card_file_box: Restructure activity, speciality and organizations
Schema refactor to correctly model the participation of an allied providing a certain speciality.

BREAKING CHANGE:
Replaced activity_allied and organization_speciality tables with an allied_participation table
with a primary key.
Changed the atention table name to "attention".
Now the attention table is related to the allied_participation table.